requirements.txt
colorama>=0.4.4
dephell-argparse>=0.1.3
pathos>=0.2.8
pyenchant>=3.2.1